% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Ar.R
\name{calculate_Ar}
\alias{calculate_Ar}
\title{Calculate Allelic Richness}
\usage{
calculate_Ar(gt)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}
}
\value{
Average allelic richness across loci.
}
\description{
This function calculates allelic richness (Ar) for each locus based on genotype data.
Allelic richness is defined as the average number of alleles per individual across all loci.
}
\details{
Reference:
- Kalinowski, S. T. (2005). https://www.montana.edu/kalinowski/software/documents/2005_HP-Rare_MolecularEcologyNotes.pdf
}
