% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Hes.R
\name{calculate_Hes}
\alias{calculate_Hes}
\title{Calculate Observed Heterozygosity (Hes)}
\usage{
calculate_Hes(gt)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}
}
\value{
Vector of observed heterozygosity values (Hes) for each locus.
}
\description{
This function calculates observed heterozygosity (Hes) for each locus based on genotype data.
Hes is the expected proportion of heterozygotes at each locus assuming HW equilibrium.
}
