% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pi_theta_d.R
\name{calculate_pi_theta_d}
\alias{calculate_pi_theta_d}
\title{Calculate Tajima's D as in Pixy}
\usage{
calculate_pi_theta_d(gt)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA. Note these methods are designed to have fixed/invariant loci                                              
included (https://onlinelibrary.wiley.com/doi/10.1111/1755-0998.13326).}
}
\description{
This function calculates Tajima's D, Watterson's theta, and pi following the methods in Pixy, which accounts for missing data to produce unbiased estimates. For further details see https://onlinelibrary.wiley.com/doi/10.1111/1755-0998.13326 and https://onlinelibrary.wiley.com/doi/10.1111/1755-0998.14104
}
