% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_private_alleles.R
\name{calculate_private_alleles}
\alias{calculate_private_alleles}
\title{Calculate Private Alleles}
\usage{
calculate_private_alleles(allele_list)
}
\arguments{
\item{allele_list}{List where each element contains a vector of alleles for a group.}
}
\value{
Data table summarizing private alleles and total alleles for each group and a total count.
}
\description{
This function calculates the number of private alleles unique to each group based on 
allele lists (`allele_list`). It computes the count of private alleles and total alleles 
for each group and aggregates results into a data table.
}
