% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rarified_Ar.R
\name{calculate_rarified_Ar}
\alias{calculate_rarified_Ar}
\title{Calculate rarified allelic richness}
\usage{
calculate_rarified_Ar(gt, site_variable)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}

\item{site_variable}{Vector of group IDs (sites) for each individual (same length as 
number of rows in gt).}
}
\value{
A data frame containing rarified allelic richness of each locus per genetic 
        group (site).
}
\description{
This function calculates rarified allelic richness across genetic groups (sites) 
using rarefaction. Rarified allelic richness estimates the average number of 
alleles in a sample while accounting for differences in sample sizes.
}
\details{
References:
- Kalinowski, S. T. (2005). https://www.montana.edu/kalinowski/software/documents/2005_HP-Rare_MolecularEcologyNotes.pdf
- Keenan, K., et al. (2013). https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210x.12067
}
