% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_total_alleles.R
\name{calculate_total_alleles}
\alias{calculate_total_alleles}
\title{Total Allele Counts}
\usage{
calculate_total_alleles(gt, group_variable, maf)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}

\item{group_variable}{Vector of genetic group IDs for each individual (same 
length as number of rows in gt). Can represent sites, populations etc.}

\item{maf}{Minimum minor allele frequency filter.}
}
\value{
Data table summarizing private alleles and total alleles for each group and a total count.
}
\description{
This function calculates the number of total and private alleles unique to each group based on 
biallelic gentoype data. It computes the count of private alleles and total alleles 
for each group and aggregates results into a data table.
}
