% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uHe.R
\name{calculate_uHe}
\alias{calculate_uHe}
\title{Calculate Unbiased Expected Heterozygosity (uHe)}
\usage{
calculate_uHe(ns, Hes)
}
\arguments{
\item{ns}{Vector containing the number of individuals sampled for each genetic group 
or locus.}

\item{Hes}{Vector containing observed heterozygosity values for each genetic group 
or locus.}
}
\value{
Average unbiased expected heterozygosity (uHe) across genetic groups or loci.
}
\description{
This function calculates unbiased expected heterozygosity (uHe) based on the number 
of individuals (ns) and observed heterozygosity (Hes) for each genetic group or locus.
Unbiased expected heterozygosity corrects for sample size biases and provides an 
estimate of heterozygosity accounting for the number of alleles and individuals.
}
