% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\docType{data}
\name{example_gt}
\alias{example_gt}
\title{Load Sample Dataset 1}
\format{
An object of class \code{data.frame} with 257 rows and 1906 columns.
}
\usage{
data(example_gt)
}
\description{
This function loads the first sample dataset stored in CSV format.
}
\examples{
data(example_gt)
head(example_gt)
}
\keyword{datasets}
