% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Load Sample Dataset 2}
\format{
An object of class \code{data.frame} with 257 rows and 3 columns.
}
\usage{
data(example_meta)
}
\description{
This function loads the second sample dataset stored in CSV format.
}
\examples{
data(example_meta)
head(example_meta)
}
\keyword{datasets}
