% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDiversity-package.R
\name{fastDiversity}
\alias{fastDiversity}
\title{fastDiversity: Population Genetics Analysis}
\description{
This package was developed to simplify population genetic analysis tasks by offering
efficient functions compatible with SNP data in biallelic format (0, 1, 2). It aims to
save time compared to other methods by providing streamlined and effective algorithms.
}
\details{
This package provides functions for basic population genetics analysis using SNP data
in biallelic format (0, 1, 2). It includes methods for calculating diversity statistics,
allelic richness, heterozygosity, and more.


The package includes functions for:
- Calculating basic diversity statistics.
- Estimating allelic richness using rarefaction.
- Identifying private alleles across populations.

For more details on each function, use `?function_name`.
}
\examples{
# Load package and data:
library(fastDiversity)
data(example_gt)
data(example_meta)

# Calculate private and total alleles by specified group (in this case, population):
allele_list_population <- make_allele_list(example_gt, example_meta$population)
ggvenn::ggvenn(allele_list_population)
private_total_alleles_population  <- calculate_private_alleles(allele_list_population)

# Calculate observed heterozygosity (Ho), expected heterozygosity (He), unbiased 
# expected heterozygosity (uHe), inbreeding coefficient (Fis), mean allelic richness (Ar),
# and rarefied allelic richness (rAr):
basicstats <- faststats(example_gt, example_meta$population,
                        example_meta$site, minimum_n=3, 
                        minimum_loci=50, maf=0.05, max_missingness=0.3)


}
\references{
Frankham, R., Ballou, J. D., & Briscoe, D. A. (2010). Introduction to Conservation Genetics (2nd ed.). Cambridge University Press. https://doi.org/10.1017/CBO9780511809002

Kalinowski ST (2005). "HP-Rare." Molecular Ecology Notes. URL: https://www.montana.edu/kalinowski/software/documents/2005_HP-Rare_MolecularEcologyNotes.pdf

Keenan K, McGinnity P, Cross TF, et al. (2013). "diveRsity: An R package for the estimation of population genetics parameters and their associated errors." Methods in Ecology and Evolution. URL: https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210x.12067
}
\seealso{
\code{\link{faststats}}, \code{\link{calculate_private_alleles}}
}
\author{
Eilish S. McMaster <emcm4052@uni.sydney.edu.au> <eilish.mcmaster@botanicgardens.nsw.gov.au>
}
\keyword{package}
\keyword{population-genetics}
