% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minor_allele_counts.R
\name{get_minor_allele_counts}
\alias{get_minor_allele_counts}
\title{Get Minor Allele Counts}
\usage{
get_minor_allele_counts(mat)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}
}
\value{
Vector of minor allele counts for each locus.
}
\description{
This function calculates the minor allele counts for each locus based on genotype data.
The minor allele count is defined as the sum of the less common allele at each locus.
}
