% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minor_allele_frequencies.R
\name{get_minor_allele_frequencies}
\alias{get_minor_allele_frequencies}
\title{Get Minor Allele Frequencies}
\usage{
get_minor_allele_frequencies(gt)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}
}
\value{
Vector of minor allele frequencies for each locus.
}
\description{
This function calculates the minor allele frequencies for each locus based on genotype data.
The minor allele frequency is defined as the frequency of the less common allele at each locus.
}
