% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_allele_list.R
\name{make_allele_list}
\alias{make_allele_list}
\title{Create Allele List by Group}
\usage{
make_allele_list(gt, site_variable)
}
\arguments{
\item{gt}{Genotype matrix where rows represent individuals and columns represent loci (0, 1, 2 format).}

\item{site_variable}{Vector specifying group IDs (sites) for each individual.}
}
\value{
List where each element contains a vector of alleles for a group.
}
\description{
This function creates a list of alleles for each group specified by `site_variable` 
based on genotype data (`gt`). It iterates over unique groups, calculates alleles 
for each group, and returns a list where each element corresponds to a group.
}
