% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_theta_d_stats.R
\name{pi_theta_d_stats}
\alias{pi_theta_d_stats}
\title{Calculate pi, theta, and D statistics}
\usage{
pi_theta_d_stats(
  gt,
  genetic_group_variable,
  site_variable,
  n = NULL,
  minimum_n = 5,
  minimum_loci = 5,
  max_missingness = 0.1,
  maf = 0,
  allele_count_min = 0
)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}

\item{genetic_group_variable}{Vector of genetic group IDs for each individual (same 
length as number of rows in gt). Can represent species, subspecies, or other 
biologically relevant structure groups.}

\item{site_variable}{Vector of group IDs (sites) for each individual (same length as 
number of rows in gt).}

\item{n}{Number of individuals to use for calculations per site. If left as NULL, 
all individuals are used.}

\item{minimum_n}{Minimum number of individuals required per site (sites with fewer 
individuals will be filtered out).}

\item{minimum_loci}{Minimum number of loci required to proceed with analysis for each 
genetic group (genetic groups with fewer loci will be skipped).}

\item{max_missingness}{Maximum proportion of missing data allowed per locus (loci with 
higher missingness will be filtered out).}

\item{maf}{Minimum minor allele frequency filter.}

\item{allele_count_min}{Minimum minor allele count to keep a locus, can be used to 
exclude singletons.}
}
\value{
A data frame with the following columns:
  \item{group}{Genetic group ID (e.g., species, subspecies).}
  \item{site}{Site ID corresponding to the sample location.}
  \item{loci}{Number of loci used in the analysis for that group-site combination.}
  \item{n}{Number of individuals sampled at the site for the given group.}
  \item{pi}{Nucleotide diversity (raw pi).}
  \item{S}{Number of segregating sites.}
  \item{theta}{Watterson's theta.}
  \item{D}{Tajima's D statistic.}
  \item{real_missingness}{Proportion of missing data in the filtered genotype matrix.}
}
\description{
This function calculates pi, Watterson's theta, and Tajima's D
for specified sites within genetic groups following the methods in Pixy,
which accounts for missing data to produce unbiased estimates. For further
details see https://onlinelibrary.wiley.com/doi/10.1111/1755-0998.13326 and
https://onlinelibrary.wiley.com/doi/10.1111/1755-0998.14104 . Data is
filtered based on genetic group (minimum number of loci, minimum minor 
allele frequency, and maximum missingness per locus) and minimum number of
individuals per site.
}
\seealso{
[resampling_individuals_CI(), resampling_loci_CI()]
}
