% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_loci_CI.R
\name{resampling_loci_CI}
\alias{resampling_loci_CI}
\title{Calculate global bootstrapped He, Ho, and FIS (parallelized) by resampling loci (as in hierfstat::boot.ppfis or arlequin)}
\usage{
resampling_loci_CI(gt, boots = 100, resample_n = NULL, CI_alpha = 0.05)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}

\item{boots}{Number of bootstrap replicates}

\item{resample_n}{Number of loci to resample with replacement -- defaults to total loci}

\item{CI_alpha}{Alpha level for confidence intervals (e.g., 0.05 for 95% CI)}
}
\value{
Named vector of confidence intervals for global FIS, HE, HO
}
\description{
Resampling loci means randomly selecting genetic markers (columns) with replacement. This keeps the same individuals, but uses different combinations of loci in each sample. It tests how your results might change depending on which loci were included.
}
