% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faststats.R
\name{faststats}
\alias{faststats}
\title{Calculate basic diversity statistics}
\usage{
faststats(
  gt,
  genetic_group_variable,
  site_variable,
  minimum_n = 3,
  minimum_loci = 50,
  maf = 0.05,
  max_missingness = 0.3
)
}
\arguments{
\item{gt}{Genotype matrix where individuals are rows and loci are columns, coded as 
0=aa, 1=aA, 2=AA.}

\item{genetic_group_variable}{Vector of genetic group IDs for each individual (same 
length as number of rows in gt). Can represent species, 
subspecies, or other biologically relevant structure groups.}

\item{site_variable}{Vector of group IDs (sites) for each individual (same length as 
number of rows in gt).}

\item{minimum_n}{Minimum number of individuals required per site (sites with fewer 
individuals will be filtered out).}

\item{minimum_loci}{Minimum number of loci required to proceed with analysis for each 
genetic group (genetic groups with fewer loci will be skipped).}

\item{maf}{Minimum minor allele frequency filter.}

\item{max_missingness}{Maximum proportion of missing data allowed per locus (loci with 
higher missingness will be filtered out).}
}
\value{
A data frame containing diversity statistics (allelic richness, heterozygosity, 
        inbreeding coefficients) per site per genetic group.
}
\description{
This function calculates allelic richness, heterozygosity, and inbreeding coefficients 
for genetic groups within specified sites. Data is filtered based on genetic group and 
site IDs, and parameters such as minimum number of individuals per site, minimum number 
of loci, minimum minor allele frequency, and maximum missingness per locus.
}
\examples{
data(example_data)
# Calculate observed heterozygosity (Ho), expected heterozygosity (He), unbiased 
# expected heterozygosity (uHe), inbreeding coefficient (Fis), mean allelic richness (Ar),
# and rarefied allelic richness (rAr):
basicstats <- faststats(example_gt, example_meta$population,
                        example_meta$site, minimum_n=3, 
                        minimum_loci=50, maf=0.05, max_missingness=0.3)
}
